Add stream handling for replicated graphs
Summary:
This fixes streams for replicated graphs where there are only
inputs and outputs, no infeed and outfeeds.

Fixes T7397

Test Plan: all tests passing, including new simple replicated graph

Reviewers: davidn, jakeh

Reviewed By: davidn

Subscribers: georgep

Maniphest Tasks: T7397

Differential Revision: https://phabricator.sourcevertex.net/D10048

+++ tensorflow/compiler/plugin/poplar/driver/executor.h
@@ -300,7 +300,7 @@ class PoplarExecutor : public se::internal::StreamExecutorInterface {
     return current_config_.retain_control_dependencies();
   }
 
-  int GetNumberOfReplicas() const {
+  int64 GetNumberOfReplicas() const {
     if (current_config_.device_config_size() > ordinal_) {
       return current_config_.device_config(ordinal_).num_replicas();
     } else {
